update QueueView layout
